vtb-ai-hr/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncio==4.0.0
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
fastapi==0.116.1
fastapi-pagination==0.14.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
minio==7.2.16
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
